phase01/W10a: fix factorial signed-overflow UB found in code review
factorial returned int, which overflows at n==13 (13! = 6'227'020'800 >
INT_MAX). Signed overflow is UB, so as a constexpr it would be a hard
compile error and at runtime the ubsan preset would abort — the exact UB
scratch.cpp demonstrates. Only caller passes 5 today, so it was latent,
but it's a trap the moment a later exercise feeds a larger value.

Changed the return type (and factorialForced) to long, which has headroom
to ~20!. The == 120 assertions are unaffected. Verified: static_assert(
factorial(13)) and factorial(20) now compile and run clean under UBSAN,
where before they were UB.

A constexpr function cannot hide UB — that's the lesson, kept in a comment.

Found by /code-review (high effort).

+++ cplus_review/phase01_modern_cpp/tests/constexpr_test.cpp
@@ -8,8 +8,13 @@
 namespace {
 
 // ============================================== constexpr: may run at compile time
-constexpr int factorial(int n) {
-    int result = 1;
+// Return type is `long`, not `int`, on purpose: int factorial overflows at n==13
+// (13! = 6'227'020'800 > INT_MAX). Signed overflow is UB, so as a constexpr it
+// would be a hard compile error, and at runtime the ubsan preset would abort —
+// the same UB scratch.cpp demonstrates. `long` buys headroom to ~20!. A constexpr
+// function cannot hide UB: that's a feature, not a limitation.
+constexpr long factorial(int n) {
+    long result = 1;
     for (int i = 2; i <= n; ++i) {
         result *= i;
     }
@@ -17,7 +22,7 @@ constexpr int factorial(int n) {
 }
 
 // consteval: MUST run at compile time
-consteval int factorialForced(int n) {
+consteval long factorialForced(int n) {
     return factorial(n);
 }
